refactor: notifications path 수정
