+++ src/components/Mobile/Notification/Notifications/components/Head/Head.jsx
@@ -1,6 +1,6 @@
 import { Stack, Typography } from "@mui/material";
 import { useState } from "react";
-import { SearchBar } from "../../common/SearchBar";
+import { SearchBar } from "../../../../common/SearchBar";
 
 export const Head = () => {
     // search
