+++ src/components/Mobile/Notification/Notifications/components/Head/index.js
