+++ src/components/Mobile/Notification/Notifications/components/NotificationView/NotificationView.jsx
@@ -1,6 +1,6 @@
 import { Stack } from "@mui/material";
-import { Head } from "../Head";
-import { NotificationTabLayout } from "../NotificationTabLayout";
+import { Head } from "../../Head";
+import { NotificationTabLayout } from "../../NotificationTabLayout";
 
 export const NotificationView = () => {
     return (
