+++ src/components/Mobile/Notification/Notifications/components/NotificationView/index.js
