+++ src/components/Mobile/Notification/Notifications/index.js
@@ -1,0 +1,1 @@
+export { NotificationView } from "./components/NotificationView";
