+++ src/components/Mobile/Notification/index.js
@@ -1,0 +1,1 @@
+export { NotificationView } from "./Notifications";
